rename folder to match component name
