+++ packages/studio-components/src/components/dropdown-list-item/dropdown-list-item.style.scss
